requirements.txt
pyowm==3.2.0
python-dotenv==0.19.0
